requirements.txt
fer==22.5.1
numpy==1.24.3
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
Pillow==11.1.0
streamlit==1.40.1
